+++ src/PipedEverything/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("PipedEverything")]
 [assembly: AssemblyProduct("PipedEverything")]
-[assembly: AssemblyVersion("1.0.24.0")]
-[assembly: AssemblyFileVersion("1.0.24.0")]
+[assembly: AssemblyVersion("1.0.25.10")]
+[assembly: AssemblyFileVersion("1.0.25.10")]